+++ include/oneapi/dpl/algorithm
@@ -27,18 +27,18 @@
 #    include "oneapi/dpl/pstl/glue_algorithm_defs.h"
 #    include "oneapi/dpl/internal/by_segment_extension_defs.h"
 #    include "oneapi/dpl/internal/binary_search_extension_defs.h"
-#    include "oneapi/dpl/internal/histogram_extension_defs.h"
+#    include "oneapi/dpl/pstl/histogram_extension_defs.h"
 #    define _ONEDPL_ALGORITHM_FORWARD_DECLARED 1
 #endif // !_ONEDPL_ALGORITHM_FORWARD_DECLARED
 
 #if _ONEDPL_EXECUTION_POLICIES_DEFINED
 // If <execution> has already been included, pull in implementations
 #    include "oneapi/dpl/pstl/glue_algorithm_impl.h"
+#    include "oneapi/dpl/pstl/histogram_impl.h"
 
 #    include "oneapi/dpl/internal/exclusive_scan_by_segment_impl.h"
 #    include "oneapi/dpl/internal/inclusive_scan_by_segment_impl.h"
 #    include "oneapi/dpl/internal/binary_search_impl.h"
-#    include "oneapi/dpl/internal/histogram_impl.h"
 #    include "oneapi/dpl/internal/reduce_by_segment_impl.h"
 #endif // _ONEDPL_EXECUTION_POLICIES_DEFINED
