+++ include/oneapi/dpl/execution
@@ -32,11 +32,11 @@
 // If the forward declarations have already been processed, pull the implementations
 #if _ONEDPL_ALGORITHM_FORWARD_DECLARED
 #    include "oneapi/dpl/pstl/glue_algorithm_impl.h"
+#    include "oneapi/dpl/pstl/histogram_impl.h"
 
 #    include "oneapi/dpl/internal/exclusive_scan_by_segment_impl.h"
 #    include "oneapi/dpl/internal/inclusive_scan_by_segment_impl.h"
 #    include "oneapi/dpl/internal/binary_search_impl.h"
-#    include "oneapi/dpl/internal/histogram_impl.h"
 #    include "oneapi/dpl/internal/reduce_by_segment_impl.h"
 #endif // _ONEDPL_ALGORITHM_FORWARD_DECLARED
